Enum anion::AnionValue

pub enum AnionValue {
    Null,
    Boolean(Option<AnionBool>),
    Integer(Option<AnionInt>),
    Float(Option<AnionFloat>),
    Decimal(Option<AnionDecimal>),
    Timestamp(Option<AnionTimestamp>),
    Str(Option<AnionString>),
    Symbol(Option<AnionSymbol>),
    Blob(Option<AnionBlob>),
    List(Option<AnionList>),
    Struct(Option<AnionStruct>),
}

Enum of all possible types of elements in an ion document.

These are mapped to either rust literal types or other anion types. All values are Options, with a None value corresponding to the equivalent 'null' value of the ion document.

Variants
Null

Pure null type

Boolean(Option<AnionBool>)

true, false, null.bool

Integer(Option<AnionInt>)

Bigint values (unlimited)

Float(Option<AnionFloat>)

64 bit floating point value

Decimal(Option<AnionDecimal>)

Exact precision real number value

Timestamp(Option<AnionTimestamp>)

Timestamp object

Str(Option<AnionString>)

String of utf8 characters

Symbol(Option<AnionSymbol>)

String of unicode characters - can be used as struct keys and in s-expressions

Blob(Option<AnionBlob>)

Raw binary data (base64 encoded)

List(Option<AnionList>)

List of AnionValues

Struct(Option<AnionStruct>)

Map of symbols to Arbitrary AnionValues

Methods

impl AnionValue

fn into_nonnull(self) -> Result<NonNullAnionValue>