Enum anion::NonNullAnionValue

pub enum NonNullAnionValue {
    Null,
    Boolean(AnionBool),
    Integer(AnionInt),
    Float(AnionFloat),
    Decimal(AnionDecimal),
    Timestamp(AnionTimestamp),
    Str(AnionString),
    Symbol(AnionSymbol),
    Blob(AnionBlob),
    List(AnionList),
    Struct(AnionStruct),
}

Variant of AnionValue enum that does not permit null values. This includes the 'pure NULL' null value, though this may be removed due to naming sillyness. This is much closer in type to true JSON values.

Variants
Null

The NULL value - in the NonNullValue!

Boolean(AnionBool)

true, false

Integer(AnionInt)

Bigint values (unlimited)

Float(AnionFloat)

64 bit floating point value

Decimal(AnionDecimal)

Exact precision real number value

Timestamp(AnionTimestamp)

Timestamp object

Str(AnionString)

String of unicode characters

Symbol(AnionSymbol)

String of unicode characters - may be used for keys

Blob(AnionBlob)

Raw binary data (base64 encoded)

List(AnionList)

List of AnionValues

Struct(AnionStruct)

Map of symbols to Arbitrary AnionValues