Function anion::from

pub fn from<T>(value: T) -> AnionValue where T: Into<AnionValue>

Return equivalent AnionValue from some object. Convenience for AnionValue::from(..).

Note: This does not parse the value, but returns an equivalent type, so anion::from("42") is an AnionString, while anion::from(42) is an AnionInt.

This function is limited by the rust type system; there is no equivalent s-expression or symbol types in rust, so those will need to be explicity cast from a string.

Example Usage

use anion;
let val = anion::from("0xf2");
assert_eq!(val, anion::AnionValue::Str(Some(anion::AnionString::from("0xf2"))));
assert_eq!(anion::from(42), anion::AnionInt::from(42));